[WIP] Remix Vite plugin support
