+++ packages/remix/package.json
@@ -4,6 +4,7 @@
   "license": "Apache-2.0",
   "main": "./dist/index.js",
   "homepage": "https://vercel.com/docs",
+  "sideEffects": false,
   "repository": {
     "type": "git",
     "url": "https://github.com/vercel/vercel.git",
